refactor : text import
