+++ src/main/java/kr/co/pinup/security/SecurityConstants.java
@@ -1,5 +1,7 @@
 package kr.co.pinup.security;
 
+import java.util.List;
+
 public class SecurityConstants {
     public static final String[] PUBLIC_URLS = {
             "/", "/images/**", "/members/login", "/api/members/oauth/**",
@@ -7,4 +9,13 @@ public class SecurityConstants {
             "/stores", "/api/stores", "/api/stores/summary",
             "/post", "/faqs"
     };
+
+    // 필터 적용 제외할 URL 목록
+    public static final List<String> EXCLUDED_URLS = List.of(
+            "/static/", "/templates/", "/css/", "/js/", "/images/", "/fonts/", "/error", "/favicon.ico",
+            "/members/login", "/api/members/oauth/",
+            "/notices", "/notices/{noticeId}", "/api/notices", "/api/notices/{noticeId}",
+            "/stores", "/api/stores", "/api/stores/summary",
+            "/post", "/faqs"
+    );
 }
