+++ src/test/java/kr/co/pinup/members/security/SecurityUtilTest.java
@@ -4,7 +4,6 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import kr.co.pinup.custom.utils.SecurityUtil;
 import kr.co.pinup.exception.common.UnauthorizedException;
 import kr.co.pinup.members.custom.WithMockMember;
 import kr.co.pinup.members.exception.OAuth2AuthenticationException;
@@ -14,6 +13,7 @@ import kr.co.pinup.oauth.OAuthProvider;
 import kr.co.pinup.oauth.OAuthService;
 import kr.co.pinup.oauth.OAuthToken;
 import kr.co.pinup.oauth.naver.NaverToken;
+import kr.co.pinup.security.SecurityUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
