+++ src/test/java/kr/co/pinup/members/security/filter/AccessTokenValidationFilterTest.java
@@ -3,10 +3,10 @@ package kr.co.pinup.members.security.filter;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import kr.co.pinup.custom.utils.SecurityUtil;
 import kr.co.pinup.members.model.dto.MemberInfo;
 import kr.co.pinup.members.security.filter.sub.TestAccessTokenValidationFilter;
 import kr.co.pinup.members.service.MemberService;
+import kr.co.pinup.security.SecurityUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
