+++ src/test/java/kr/co/pinup/members/service/MemberServiceTest.java
@@ -1,6 +1,5 @@
 package kr.co.pinup.members.service;
 
-import kr.co.pinup.custom.utils.SecurityUtil;
 import kr.co.pinup.exception.common.UnauthorizedException;
 import kr.co.pinup.members.Member;
 import kr.co.pinup.members.custom.MemberTestAnnotation;
@@ -13,6 +12,7 @@ import kr.co.pinup.members.model.dto.MemberResponse;
 import kr.co.pinup.members.model.enums.MemberRole;
 import kr.co.pinup.members.repository.MemberRepository;
 import kr.co.pinup.oauth.OAuthProvider;
+import kr.co.pinup.security.SecurityUtil;
 import org.junit.jupiter.api.*;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
